pyproject.toml
[project]
name = "code-courtroom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.2.1",
    "pydantic>=2.11.5",
    "ruff>=0.11.13",
]
